Still working out radix sort,
Basically, trying to figure out what to do if there are missing digits.
If a missing digits always counts as less than a present digit, then you
sort have this as sorted: ['2', '3', '11']. If you say that the missing
digit is always greater then you have this as sorted: ['11', '1', '2'].

From a little Google research it appears that other people haven't
worked this out and implement +/- the same way as it is done here
(decide to either pad numbers with fewer digits with zeros for example).
If there were some test cases then that would make things a lot easier.
It is difficult to both spec and implement.